Remove _helpers export, move GridPropHelpers
