+++ packages/react-data-grid/src/__tests__/GridPropHelpers.ts
@@ -1,4 +1,4 @@
-import { CalculatedColumn, CellMetaData } from '../../common/types';
+import { CalculatedColumn, CellMetaData } from '../common/types';
 
 export interface Row {
   id?: number;
