--- packages/react-data-grid/src/helpers/index.ts
@@ -1,3 +1,0 @@
-import GridPropHelpers from './test/GridPropHelpers';
-
-export const test = { GridPropHelpers };
