+++ packages/react-data-grid/src/index.ts
@@ -2,12 +2,8 @@ export { default, ReactDataGridProps } from './ReactDataGrid';
 export { default as Cell } from './Cell';
 export { default as Row } from './Row';
 export { default as HeaderCell } from './HeaderCell';
-import * as _helpers from './helpers';
 export * from './Columns';
 export * from './formatters';
 export * from './common/editors';
 export * from './common/enums';
 export * from './common/types';
-export {
-  _helpers
-};
